refactor: run script will allow for config and dataset options
This is part of the SET-490 ticket where we need a good example of
multicohorts with an analysis_dataset that differs from the child
cohort datasets. So creating a new config for this test case. That
requires the option to run with a different config in our little
run-test-workflow.sh helper script. As well as the new config.

SET-490